// Bullhorn-Attachments-UI-Code/src/app/shared/models/Data-Table/worker-search-data-grid/assignment-information-tbl.model.ts
export interface AssignmentInformationTable {
    hcmAssignmentId?:  [string | null  | undefined];
    fusionCustomerAccountName?:  [string | null  | undefined];
    project?:  [string | null  | undefined];
    jobTitle?:  [string | null  | undefined];
    assignmentStatus?:  [string | null  | undefined];
    actualStartDate?:  [string | null  | undefined];
    actualEndDate?:  [string | null  | undefined];
    targetStartDate?:  [string | null  | undefined];
    targetEndDate?:  [string | null  | undefined];
    suspendDate?:  [string | null  | undefined];
    plannedEndDate?:  [string | null  | undefined];
    standardPayRate?:  [string | null  | undefined];
    standardBillRate?:  [string | null  | undefined];
    workState?:  [string | null  | undefined];
    createdBy?:[string | null | undefined];
    createdOn?:[string | null | undefined];
    updatedBy?:[string | null | undefined];
    updatedOn?:[string | null | undefined];
    assignmentUnqId?:  [number | null  | undefined];
}